+++ Tests/Invoke-AlerterNotification.tests.ps1
@@ -8,7 +8,7 @@ Get-Module $Module | Remove-Module -Force
 Import-Module "$Root/$Module" -Force
 
 Describe "Invoke-AlerterNotification PS$PSVersion" {
-    
+
     InModuleScope MacNotify {
         Mock Invoke-Alerter {}
 
@@ -24,5 +24,5 @@ Describe "Invoke-AlerterNotification PS$PSVersion" {
             Invoke-AlerterNotification -Message 'How are you' -Subtitle 'Test Subtitle' | Should -Be $null
             Assert-MockCalled Invoke-Alerter -Times 1 -Exactly -Scope It
         }
-    }   
+    }
 }