+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## !Deploy
+
+* Fixed Invoke-MacNotification and Invoke-AlertNotification to handle single or double quotes in the message, title or subtitle. Fixes #3.
+
 ## [1.1.29] - 2024-03-17
 
 This version of MacNotify has been tested under MacOS 14 and works. It may not work in earlier versions. Version 1.1.27 was tested and working in MacOS 10.